chore(batch-exports): move the Redshift host test to its destination file
Generated-By: PostHog Desktop
Task-Id: 25c2e0bb-7208-4c1a-bbae-c2c528424ea3
