+++ products/batch_exports/backend/tests/api/test_create.py
@@ -6,7 +6,6 @@ from zoneinfo import ZoneInfo
 import pytest
 from unittest import mock
 
-from django.test import override_settings
 from django.test.client import Client as HttpClient
 
 from asgiref.sync import async_to_sync
@@ -772,53 +771,3 @@ def test_creating_batch_export_with_filters(
 
     if expected_error:
         assert expected_error in response.json()["detail"]
-
-
-@pytest.mark.parametrize(
-    "host",
-    [
-        "192.168.1.1",
-        "127.0.0.1",
-        "[::1]",
-        "10.0.0.1",
-        "169.254.0.0",
-        "localhost",
-    ],
-)
-def test_create_redshift_batch_export_fails_with_invalid_host(
-    client: HttpClient, temporal, organization, team, user, host, aws_redshift_integration
-):
-    """Test creating a BatchExport with Redshift destination validates inputs for 'COPY'.
-
-    Postgres host validation is covered separately in test_create_postgres.py, where the host
-    comes from the linked Integration rather than from inline config.
-    """
-
-    destination_data = {
-        "type": "Redshift",
-        "config": {
-            "database": "my-db",
-            "host": host,
-            "schema": "public",
-            "table_name": "my_events",
-        },
-        "integration": aws_redshift_integration.pk,
-    }
-
-    batch_export_data = {
-        "name": "my-production-destination",
-        "destination": destination_data,
-        "interval": "hour",
-    }
-
-    client.force_login(user)
-
-    with override_settings(TEST=0, DEBUG=0):
-        response = create_batch_export(
-            client,
-            team.pk,
-            batch_export_data,
-        )
-
-    assert response.status_code == status.HTTP_400_BAD_REQUEST, response.json()
-    assert f"Invalid host: '{host}'" in response.json()["detail"]
